html/semantics/sections/the-body-element/sections-body-events-dom-content.html

<!DOCTYPE html>
<html>
<head>
    <meta charset="utf-8">
    <title>Body Element</title>
    <link rel="author" title="dzenana" href="mailto:dzenana.trenutak@gmail.com">
    <link rel="help" href="http://www.w3.org/html/wg/drafts/html/CR/single-page.html#the-body-element">
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>
<body
  onblur="myReaction = 'blur'"
  onerror="myReaction = 'error'"
  onfocus="myReaction = 'focus'"
  onload="myReaction = 'load'"
  onresize="myReaction = 'resize'"
  onscroll="myReaction = 'scroll'"

  onafterprint="myReaction = 'afterprint'"
  onbeforeprint="myReaction = 'beforeprint'"
  onbeforeunload="myReaction = 'beforeunload'"
  onhashchange="myReaction = 'hashchange'"
  onlanguagechange="myReaction = 'languagechange'"
  onmessage="myReaction = 'message'"
  onoffline="myReaction = 'offline'"
  ononline="myReaction = 'online'"
  onpagehide="myReaction = 'pagehide'"
  onpageshow="myReaction = 'pageshow'"
  onpopstate="myReaction = 'popstate'"
  onstorage="myReaction = 'storage'"
  onunload="myReaction = 'unload'"
>

    <h1>Description</h1>
    <p>This test validates that all automatable body element content attribute event handlers react to relevant events.</p>
    <p>Please be patient - there is a 10 second delay between events being fired and the tests being run.</p>
    <p>For this round of tests, events are fired via DOM createEvent and dispatchEvent methods.</p>

    <div id="log"></div>

    <script>
        "use strict";

        // Verify that all the content attributes that allow body to react to certain events actually do react to those events

        setup({ explicit_done: true, allow_uncaught_exception: true });

        var myReaction = "";

        var windowEvents = [
            // Table 2.
            "blur",
            "error",
            "focus",
            "load",
            "resize",
            "scroll",

            // Table 3.
            "afterprint",
            "beforeprint",
            "beforeunload",
            "hashchange",
            "languagechange",
            "message",
            "offline",
            "online",
            "pagehide",
            "pageshow",
            "popstate",
            "storage",
            "unload",
        ];
        windowEvents.forEach(function(type) {
            test(function() {
                assert_true(document.body.hasAttribute("on" + type),
                            "Should have the attribute set on the body.");

                myReaction = "";
                window.dispatchEvent(new Event(type));
                assert_equals(myReaction, type,
                              "Event handler should have been triggered");
            }, "Body.on" + type + " (fired on window)");

            test(function() {
                assert_true(document.body.hasAttribute("on" + type),
                            "Should have the attribute set on the body.");

                myReaction = "";
                document.body.dispatchEvent(new Event(type));
                assert_equals(myReaction, "",
                              "Event handler should not have been triggered");
            }, "Body.on" + type + " (fired on body)");
        });
        done();
    </script>

</body>
